Refactor InspectionFormDialog component
Refactor InspectionFormDialog.tsx into smaller, more manageable components to improve maintainability.
