+++ src/components/residence/inspection/InspectionFormDialog.tsx
@@ -1,5 +1,4 @@
 
-import { useState } from "react";
 import {
   Dialog,
   DialogContent,
@@ -9,12 +8,11 @@ import {
   DialogDescription,
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
-import { InspectionRoom } from "./InspectionRoom";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import type { Room } from "@/types/api";
 import type { InspectionRoom as InspectionRoomType } from "./types";
 import { InspectionInfoStep } from "./form/InspectionInfoStep";
-import { BasicInformation } from "./form/BasicInformation";
+import { InspectionTabs } from "./form/InspectionTabs";
+import { useInspectionForm } from "@/hooks/useInspectionForm";
 
 interface InspectionFormDialogProps {
   isOpen: boolean;
@@ -23,120 +21,21 @@ interface InspectionFormDialogProps {
   rooms: Room[];
 }
 
-const initialRoomData: InspectionRoomType = {
-  roomId: "",
-  conditions: {
-    wall1: "",
-    wall2: "",
-    wall3: "",
-    wall4: "",
-    floor: "",
-    ceiling: "",
-    details: ""
-  },
-  actions: {
-    wall1: [],
-    wall2: [],
-    wall3: [],
-    wall4: [],
-    floor: [],
-    ceiling: [],
-    details: []
-  },
-  componentNotes: {
-    wall1: "",
-    wall2: "",
-    wall3: "",
-    wall4: "",
-    floor: "",
-    ceiling: "",
-    details: ""
-  },
-  photos: [],
-  isApproved: false,
-  isHandled: false
-};
-
 export function InspectionFormDialog({ isOpen, onClose, onSubmit, rooms }: InspectionFormDialogProps) {
-  const [inspectorName, setInspectorName] = useState("");
-  const [step, setStep] = useState<"info" | "inspection">("info");
-  const [expandedRoomIds, setExpandedRoomIds] = useState<string[]>([]);
-  const [inspectionData, setInspectionData] = useState<Record<string, InspectionRoomType>>(() => {
-    const initialData: Record<string, InspectionRoomType> = {};
-    rooms.forEach(room => {
-      initialData[room.id] = {
-        ...initialRoomData,
-        roomId: room.id,
-      };
-    });
-    return initialData;
-  });
-
-  const handleNext = (e: React.FormEvent) => {
-    e.preventDefault();
-    setStep("inspection");
-    if (rooms.length > 0) {
-      setExpandedRoomIds([rooms[0].id]);
-    }
-  };
-
-  const handleConditionUpdate = (
-    roomId: string,
-    field: keyof InspectionRoomType["conditions"],
-    value: string
-  ) => {
-    setInspectionData(prev => ({
-      ...prev,
-      [roomId]: {
-        ...prev[roomId],
-        conditions: {
-          ...prev[roomId].conditions,
-          [field]: value
-        }
-      }
-    }));
-  };
-
-  const handleActionUpdate = (
-    roomId: string,
-    field: keyof InspectionRoomType["actions"],
-    action: string
-  ) => {
-    setInspectionData(prev => {
-      const currentActions = prev[roomId].actions[field];
-      const newActions = currentActions.includes(action)
-        ? currentActions.filter(a => a !== action)
-        : [...currentActions, action];
-
-      return {
-        ...prev,
-        [roomId]: {
-          ...prev[roomId],
-          actions: {
-            ...prev[roomId].actions,
-            [field]: newActions
-          }
-        }
-      };
-    });
-  };
-
-  const handleComponentNoteUpdate = (
-    roomId: string,
-    field: keyof InspectionRoomType["componentNotes"],
-    note: string
-  ) => {
-    setInspectionData(prev => ({
-      ...prev,
-      [roomId]: {
-        ...prev[roomId],
-        componentNotes: {
-          ...prev[roomId].componentNotes,
-          [field]: note
-        }
-      }
-    }));
-  };
+  const {
+    inspectorName,
+    setInspectorName,
+    step,
+    setStep,
+    expandedRoomIds,
+    inspectionData,
+    handleNext,
+    handleCancel: resetForm,
+    handleToggleRoom,
+    handleConditionUpdate,
+    handleActionUpdate,
+    handleComponentNoteUpdate
+  } = useInspectionForm(rooms);
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
@@ -146,24 +45,13 @@ export function InspectionFormDialog({ isOpen, onClose, onSubmit, rooms }: Inspe
 
   const handleCancel = () => {
     onClose();
-    setStep("info");
-    setInspectorName("");
-    setExpandedRoomIds([]);
+    resetForm();
   };
 
   const handleOpenChange = (open: boolean) => {
     if (!open) {
       handleCancel();
     }
-  };
-
-  const handleToggleRoom = (roomId: string) => {
-    setExpandedRoomIds(prev => {
-      if (prev.includes(roomId)) {
-        return prev.filter(id => id !== roomId);
-      }
-      return [...prev, roomId];
-    });
   };
 
   return (
@@ -191,55 +79,16 @@ export function InspectionFormDialog({ isOpen, onClose, onSubmit, rooms }: Inspe
         ) : (
           <form onSubmit={handleSubmit}>
             <div className="space-y-4 py-4">
-              <Tabs defaultValue="basic" className="w-full">
-                <TabsList className="w-full justify-start bg-background border-b rounded-none px-0 overflow-x-auto flex-nowrap">
-                  <TabsTrigger value="basic" className="text-sm sm:text-base whitespace-nowrap data-[state=active]:border-b-2 data-[state=active]:border-primary rounded-none">
-                    Grundläggande info
-                  </TabsTrigger>
-                  <TabsTrigger value="protocol" className="text-sm sm:text-base whitespace-nowrap data-[state=active]:border-b-2 data-[state=active]:border-primary rounded-none">
-                    Protokoll
-                  </TabsTrigger>
-                  <TabsTrigger value="floorplan" className="text-sm sm:text-base whitespace-nowrap data-[state=active]:border-b-2 data-[state=active]:border-primary rounded-none">
-                    Planritning
-                  </TabsTrigger>
-                </TabsList>
-
-                <TabsContent value="basic" className="mt-6">
-                  <BasicInformation
-                    inspectorName={inspectorName}
-                    roomCount={rooms.length}
-                  />
-                </TabsContent>
-
-                <TabsContent value="protocol" className="mt-6">
-                  <div className="space-y-4">
-                    {rooms.map(room => (
-                      <InspectionRoom
-                        key={room.id}
-                        room={room}
-                        isExpanded={expandedRoomIds.includes(room.id)}
-                        onToggle={() => handleToggleRoom(room.id)}
-                        inspectionData={inspectionData[room.id]}
-                        onConditionUpdate={(field, value) => 
-                          handleConditionUpdate(room.id, field, value)
-                        }
-                        onActionUpdate={(field, action) => 
-                          handleActionUpdate(room.id, field, action)
-                        }
-                        onComponentNoteUpdate={(field, note) => 
-                          handleComponentNoteUpdate(room.id, field, note)
-                        }
-                      />
-                    ))}
-                  </div>
-                </TabsContent>
-
-                <TabsContent value="floorplan" className="mt-6">
-                  <div className="flex items-center justify-center h-[400px] border-2 border-dashed rounded-lg">
-                    <p className="text-muted-foreground">Planritning är inte tillgänglig</p>
-                  </div>
-                </TabsContent>
-              </Tabs>
+              <InspectionTabs
+                inspectorName={inspectorName}
+                rooms={rooms}
+                expandedRoomIds={expandedRoomIds}
+                inspectionData={inspectionData}
+                onToggleRoom={handleToggleRoom}
+                onConditionUpdate={handleConditionUpdate}
+                onActionUpdate={handleActionUpdate}
+                onComponentNoteUpdate={handleComponentNoteUpdate}
+              />
             </div>
 
             <DialogFooter className="flex-col-reverse sm:flex-row gap-2">
