+++ src/components/residence/inspection/form/initialData.ts
@@ -1,0 +1,47 @@
+
+import type { InspectionRoom } from "../types";
+
+export const initialRoomData: InspectionRoom = {
+  roomId: "",
+  conditions: {
+    wall1: "",
+    wall2: "",
+    wall3: "",
+    wall4: "",
+    floor: "",
+    ceiling: "",
+    details: ""
+  },
+  actions: {
+    wall1: [],
+    wall2: [],
+    wall3: [],
+    wall4: [],
+    floor: [],
+    ceiling: [],
+    details: []
+  },
+  componentNotes: {
+    wall1: "",
+    wall2: "",
+    wall3: "",
+    wall4: "",
+    floor: "",
+    ceiling: "",
+    details: ""
+  },
+  photos: [],
+  isApproved: false,
+  isHandled: false
+};
+
+export const initializeInspectionData = (rooms: { id: string }[]) => {
+  const initialData: Record<string, InspectionRoom> = {};
+  rooms.forEach(room => {
+    initialData[room.id] = {
+      ...initialRoomData,
+      roomId: room.id,
+    };
+  });
+  return initialData;
+};
